requirements.txt
Flask==1.1.2
Pillow==8.1.0
requests==2.25.0
pip==20.2.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7